+++ NEWS.md
@@ -8,7 +8,7 @@
 
 ### New & amended functions
 * `dplyr::distinct()` can be used to find grouped data and summaries, generalising `atlas_species()`
-* new functions `as_query()` and `coalesce()` as prequels to `collapse()`
+* new functions `as_query()` and `compound()` as prequels to `collapse()`
 * `galah_call()` is now synonmous with `request_data()` rather than wrapping all `request_` functions; `method` argument is removed.
 
 ## Changes to metadata functions